docs(ui): Убраны лишние stories, декоратор Button stories, theming & colors
