+++ packages/ui/src/components/Colors.stories.mdx
@@ -1,9 +1,8 @@
 import { Meta, Props, Canvas, Story, Description } from '@storybook/addon-docs/blocks';
 
-import * as stories from './Palette.stories';
+import * as stories from './Colors.stories';
 
 <Meta title="Colors" />
-
 
 # Цветовая палитра
