+++ packages/ui/src/components/Skeleton/Skeleton.stories.tsx
@@ -82,3 +82,9 @@ export const InCard = () => {
         </Card>
     );
 };
+
+InCard.parameters = {
+    chromatic: {
+        disable: true,
+    },
+};
