+++ packages/ui/src/components/Stepper/Stepper.stories.mdx
@@ -61,5 +61,5 @@ export const App = () => (
 
 ### Использование составных компонентов
 <Canvas>
-    <Story name="Custom" story={stories.Custom} />
+    <Story name="CustomAssembly" story={stories.CustomAssembly} />
 </Canvas>
