--- packages/ui/src/components/pallete.mdx
@@ -1,26 +1,0 @@
----
-name: Цветовая палитра
-route: /pallete
----
-
-import { Playground } from 'docz'
-
-
-import { ThemeSelect } from '../helpers/ThemeSelect'
-import { ThemeBackground } from '../helpers/ThemeBackground'
-import { ColorsDemo } from '../helpers/ColorsDemo';
-
-
-<ThemeSelect />
-
-# Цветовая палитра
-
-
-<div style={{ margin: 20, marginTop: 60 }} >
-  <ThemeBackground>
-      <div style={{ margin: 20 }}>
-        <ColorsDemo />
-      </div>
-  </ThemeBackground>
-</div>
-
