+++ packages/ui/src/helpers/Palette.tsx
@@ -6,7 +6,44 @@ import { Cell, CellIcon } from '../components/Cell';
 import { TextBox } from '../components/TextBox';
 import { Display3, Headline3, Footnote1, Caption } from '../components/Typography';
 
-import { themeColorsComments } from './ColorsDemo';
+export const themeColorsComments = {
+    accent: 'Акцентный цвет призыва к действию',
+    gradient: 'Градиент для заливки основного фона',
+
+    text: 'Базовый цвет текста, совпадает с primary',
+    primary: 'Первичный цвет текста',
+    secondary: 'Вторичный цвет текста',
+    tertiary: 'Третичный цвет текста',
+
+    background: 'Базовый цвет фона, совпадает с backgroundPrimary',
+    backgroundPrimary: 'Первичный цвет фона',
+    backgroundSecondary: 'Вторичный цвет фона',
+    backgroundTertiary: 'Третичный цвет текста',
+
+    warning: 'Цвет предупреждения',
+    critical: 'Цвет ошибки',
+
+    overlay: 'Цвет фона паранжи',
+
+    surfaceLiquid01: 'Цвет подложки 1',
+    surfaceLiquid02: 'Цвет подложки 2',
+    surfaceLiquid03: 'Цвет подложки 3',
+    surfaceCard: 'Цвет подложки карточек',
+
+    buttonPrimary: 'Первичный цвет контролов',
+    buttonSecondary: 'Вторичный цвет контролов',
+
+    buttonAccent: 'Акцентный цвет у контролов',
+    buttonWarning: 'Цвет предупреждения у контролов',
+    buttonCritical: 'Цвет ошибки у контролов',
+    buttonChecked: 'Цвет зажатого контрола',
+    buttonFocused: 'Цвет рамки фокуса у контрола',
+
+    speechBubbleSent: 'Цвет фона баблов отправленный сообщений',
+    speechBubbleReceived: 'Цвет фона баблов получнных сообщений',
+
+    voicePhraseGradient: 'Градиент подсказок о голосовых запросах',
+};
 
 const themes = {
     darkEva,
@@ -51,7 +88,7 @@ export interface PaletteProps {
     heading?: string;
 }
 
-export const Palette: React.FC<PaletteProps> = ({ theme, title, heading }) => {
+export const Palette: React.FC<PaletteProps> = ({ theme = 'darkSber', title, heading }) => {
     const selectedTheme = themes[theme][':root'];
     const primary = selectedTheme['--plasma-colors-primary'];
     const items = React.useMemo(() => {
@@ -65,7 +102,7 @@ export const Palette: React.FC<PaletteProps> = ({ theme, title, heading }) => {
             {heading && <StyledHeading color={primary}>{heading}</StyledHeading>}
             {items.map(([key, value]) => {
                 const prefix = '--plasma-colors-';
-                const name = key.replace(prefix, '');
+                const name = key.replace(prefix, '') as keyof typeof themeColorsComments;
                 return (
                     <Cell
                         key={name}
