+++ packages/ui/src/helpers/StoryDecorators.tsx
@@ -5,20 +5,20 @@ import { Container } from '../components/Grid';
 
 import { GridLines } from './GridLines';
 
-export const InSpacing = (Story) => (
+export const InSpacing = (Story: React.FC) => (
     <div style={{ padding: '1rem' }}>
         <Story />
     </div>
 );
 
-export const WithGridLines = (Story) => (
+export const WithGridLines = (Story: React.FC) => (
     <>
         {boolean('Display grid', true) && <GridLines />}
         <Story />
     </>
 );
 
-export const InContainer = (Story) => (
+export const InContainer = (Story: React.FC) => (
     <Container>
         <Story />
     </Container>
